Move most property filter methods from PromptRunner to ObjectCreator (#1160)
This commit moves the withProperties and withoutProperties methods from
PromptRunner to ObjectCreator, including vararg and KProperty variants.

This move makes the PromptRunner cleaner, and ensures that the various
filtering methods are offered in the most suitable context.

The main filter method withPropertyFilter exists on both ObjectCreator
and PromptRunner, as the default implementation of the former
(PromptRunnerObjectCreator) uses the latter as a delegate.

Signed-off-by: Arjen Poutsma <poutsma@mac.com>